+++ Changelog.md
@@ -1,7 +1,8 @@
 ### 1.2.0
 
 * Make it easy to include `negate` and `matcher_only` matchers.
-* Move RSpec::Matchers::ChangeToNow.override_to setting to RSpec::ChangeToNow.override_to
+* Move `RSpec::Matchers::ChangeToNow.override_to` setting to `RSpec::ChangeToNow.override_to`
+* When passed both arguments and a block, `detect` will raise a `SyntaxError` instead of `ExpectationNotMet`
 
 ### 1.1.0